Finished the MathClient Task
The important thing to notice in this task is that there is a set pattern. We can therefore use that to our advantage. The Classes used to decipher what our programm would do with the cmd that it recived was: Pattern and the String method .subString(). As a regex to decipher if the string contained letter I used ".*[a-zA-Z].*".