Finished the chat with textUI
In this push I am adding four new classes. These are all for the chat assignment. The classes are ChatClient, ChatServer, ChatClientThread and ChatServerThread. The whole idea is that the client sends the information trough both the thread where the serverThread can communicate with the server to tell all the other threads that one of the client have made an input.